Conflicts:
	all/build.xml
	fs/descriptors/org.jnode.partitions.apm.xml
	fs/src/fs/org/jnode/fs/ext2/Ext2FileSystem.java
	fs/src/fs/org/jnode/fs/ext2/INodeTable.java
	fs/src/fs/org/jnode/fs/hfsplus/HfsPlusFile.java
	fs/src/fs/org/jnode/fs/hfsplus/catalog/Catalog.java
	fs/src/fs/org/jnode/fs/hfsplus/catalog/CatalogFolder.java
	fs/src/fs/org/jnode/fs/hfsplus/catalog/CatalogKey.java
	fs/src/fs/org/jnode/fs/hfsplus/tree/BTHeaderRecord.java
	fs/src/fs/org/jnode/fs/jfat/FatEntry.java
	fs/src/fs/org/jnode/partitions/gpt/GptPartitionTable.java
	fs/src/fs/org/jnode/partitions/gpt/GptPartitionTableEntry.java
	fs/src/fs/org/jnode/partitions/gpt/GptPartitionTableType.java
	fs/src/fs/org/jnode/partitions/gpt/GptPartitionTypes.java
	fs/src/test/org/jnode/partitions/ibm/IBMPartitionTableEntryTest.java

+++ cli/src/commands/org/jnode/command/common/ExprCommand.java
@@ -1,7 +1,7 @@
 /*
- * $Id$
+ * $Id: header.txt 5714 2010-01-03 13:33:07Z lsantha $
  *
- * Copyright (C) 2003-2010 JNode.org
+ * Copyright (C) 2003-2012 JNode.org
  *
  * This library is free software; you can redistribute it and/or modify it
  * under the terms of the GNU Lesser General Public License as published
@@ -24,7 +24,6 @@ import java.io.PrintWriter;
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.Set;
-
 import org.jnode.shell.AbstractCommand;
 
 
@@ -319,7 +318,7 @@ public class ExprCommand extends AbstractCommand {
     
     private int asInteger(Object obj) throws ExprException {
         if (obj instanceof Integer) {
-            return ((Integer) obj).intValue();
+            return (Integer) obj;
         } else {
             try {
                 return Integer.parseInt(obj.toString());
